Starts all downloads, displays download links
Downloads are now started for every single song in a playlist and a
button to download the whole playlist is immediately displayed and will
work off the bat, however for it to be all the songs you want and to be
completely downloaded you have to wait like 15 seconds per song but a
user doesn't know that.
Will make a script to check when the playlist is done and display the
link then.